pyproject.toml
[project]
name = "qdrant-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "fastembed>=0.7.1",
    "llama-index>=0.13.0",
    "llama-index-embeddings-fastembed>=0.4.0",
    "llama-index-llms-openai>=0.5.0",
    "llama-index-readers-file>=0.5.0",
    "llama-index-vector-stores-qdrant>=0.7.1",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
]
